improve argus code quality (#1582)
1) fix warnings: -Werror (#1582)
2) 修复部分用例对路径有依赖的问题。
3) 适配gcc 9.3.1部分特性(return rvalue)

+++ core/cms/cloudMonitor/metric_calculate_2.h
@@ -95,7 +95,7 @@ namespace cloudMonitor {
             return ret;
         }
 
-        unsigned int Count() const {
+        size_t Count() const {
             return mCount;
         }
 
@@ -114,7 +114,7 @@ namespace cloudMonitor {
         TMetric mMin;
         TMetric mTotal;
         TMetric mLast;
-        unsigned int mCount = 0;
+        size_t mCount = 0;
     };
 }
 #endif //ARGUSAGENT_METRIC_CALCULATE_2_H
